requirements.txt
autoflake==2.3.1
black==24.8.0
pre-commit==3.8.0
click==8.1.7
isort==5.13.2
flake8==7.1.1
flake8-formatter-junit-xml==0.0.6
#
setuptools==75.0.0
Django==5.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
django-cors-headers==4.4.0
parameterized==0.9.0
freezegun==1.5.1
vk-api==11.9.9
beautifulsoup4==4.12.3
requests==2.32.3
gigachat==0.1.35
ics==0.7.2
nltk==3.9.1
pymorphy3==2.0.2
social-auth-app-django
python-social-auth[django]
django-rest-framework-social-oauth2
django-oauth-toolkit
openpyxl==3.1.5
